pyproject.toml
[tool.poetry]
name = "SGP"
version = "0.1.0"
description = ""
authors = ["cesar.moreira.oliveira <cesardddp@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask-SQLAlchemy = "^2.4.4"
SQLAlchemy = "^1.3.22"
sqlalchemy-migrate = "^0.13.0"
Flask-Migrate = "^2.5.3"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.1"
pytest-flask = "^1.1.0"
Flask-Reuploaded = "^0.5.0"
Flask-Cors = "^3.0.10"
Flask = "^2.0.1"
Flask-Login = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
ipython = "^7.19.0"
flask-shell-ipython = "^0.4.1"
ipdb = "^0.13.9"
pandas = "^1.3.3"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
